add: Notification component
Add Notification component and use it in ExplorePage

The commit adds a new Notification component to the basic_elements folder. The Notification component is then imported into the ExplorePage.svelte file, where it is used to display an error message when authentication fails. The package version number has also been updated from 0.11.3 to 0.12.0 in both package-lock.json and package.json files.