add: cover button to snap images
This commit adds a new feature that allows users to set a cover image for snaps. It includes changes to the Button component and the addition of a new Image component. The Image component displays all available images and provides an option to set one as the cover image by clicking on the "Cover" button. When hovering over an image, if it is already set as the cover, a disabled "Cover" button will appear; otherwise, an active "Cover" button will appear.