add: ImagesEmpty component and use it in SnapUpsertPage
This commit adds a new component called ImagesEmpty which displays a message when there are no images to show. The SnapUpsertPage now uses this component to display the message "Please add images" when there are no images available. The lodash library is also updated to import only get, set, and isEmpty functions instead of the entire library.