add: ComingSoon component to ProjectPage
This commit adds a new ComingSoon component to the ProjectPage. The component is displayed when the Recs tab is selected and no recommendations are found. The commit also includes changes to import and export statements in index.js, as well as styling updates for snaps_layout and coming_soon_layout classes. Finally, this commit includes a new story for the ProjectPage with the Recs tab selected.