refactor: clickProject function and Input component in SnapUpsertActions.svelte
The clickProject function in ProjectCard.svelte was refactored to remove a console.log statement. The Input component in SnapUpsertActions.svelte was also refactored to use the bind directive for value and add an object with min and max length properties. Additionally, handlePublish function in SnapUpsertPage.svelte now logs the snap_name property from the event detail object.