Add error message when images are empty
This commit adds an error message to the "Add Images" button in SnapUpsertActions.svelte component. The error message is displayed when there are no images added and the user tries to publish a snap. The AddImagesButton.svelte component now exports a new prop called "images_empty_error" which is used to conditionally display the error message.