refactor: components
The package version has been incremented. Significant refactoring was done on the Feedback.svelte, TopicSidebar.svelte, ProjectPage.stories.svelte components. The changes include modifying how topics are handled in the Feedback component and adding a condition to handle zero topics in the TopicSidebar component. A new story for ProjectPageFeedback was also added.