refactor: UI components
The minor version of the package has been incremented. In terms of code changes, unnecessary calculations for action bar height in Conversation component have been removed to simplify the code. The CSS styling for content class in the same component has also been adjusted to improve layout.

In addition, a conditional rendering check has been added to Feedback component to ensure that Conversation component is only rendered when there are topics available. This prevents potential errors or empty states.